Distribute item_immutable store to verifiers

+++ src/main/java/net/jackw/olep/message/TestMessage.java
@@ -6,7 +6,7 @@ import com.google.errorprone.annotations.Immutable;
 @Immutable
 public class TestMessage extends TransactionRequestBody {
     public final String body;
-    public final long item;
+    public final int item;
 
     public TestMessage(@JsonProperty("body") String body, @JsonProperty("item") int item) {
         this.body = body;
